[P3][refactor] enhanced-resources: Refactor MarbleDataAccess (CAP-028)
Refactorings applied:
- Remove 19 #region blocks obscuring code structure
- Change class visibility from public to internal (codebase pattern)
- Extract ImageEntry record to own file (one-type-per-file pattern)
- Consolidate 12 per-method provenance blocks into single class-level remarks
- Add XML doc comments to all public methods with thread safety notes
- Simplify ImageOverlapsRange by removing redundant intermediate variables
- Apply expression-bodied members for short methods
- Simplify HaveMarbleData LINQ to use method group reference

File reduction: 522 -> 387 lines (MarbleDataAccess.cs) + 17 lines (ImageEntry.cs)

Tests: All 56 passing (698 full suite)
Agent: tdd-refactorer
